+++ docs/make.jl
@@ -1,11 +1,11 @@
-using Documenter, CABLAB, CABLABPlots
+using Documenter, ESDL, ESDLPlots
 
 
 makedocs(
-    modules = [CABLAB],
+    modules = [ESDL],
     clean   = false,
     format   = :html,
-    sitename = "CABLAB.jl",
+    sitename = "ESDL.jl",
     authors = "Fabian Gans",
     pages    = Any[ # Compat: `Any` for 0.4 compat
         "Home" => "index.md",
@@ -22,7 +22,7 @@ makedocs(
 
 deploydocs(
     #deps   = Deps.pip("mkdocs", "python-markdown-math"),
-    repo   = "github.com/CAB-LAB/CABLAB.jl.git",
+    repo   = "github.com/CAB-LAB/ESDL.jl.git",
     julia  = "0.5",
     deps   = nothing,
     make   = nothing,