+++ src/Proc/Proc.jl
@@ -1,7 +1,7 @@
 module Proc
 export DAT_detectAnomalies!, removeMSC, gapFillMSC, normalizeTS, simpleAnomalies,
   sampleLandPoints, toPointAxis, getMSC, filterTSFFT, getNpY,timespacequantiles,timelonlatquantiles, getMedSC,DATfitOnline,
-  extractLonLats, cubePCA, rotation_matrix, transformPCA, explained_variance
+  extractLonLats, cubePCA, rotation_matrix, transformPCA, explained_variance,exportcube
 importall ..DAT, ..CubeAPI, ..Cubes
 macro no_ocean(maskin,maskout)
     esc(quote