+++ test/runtests.jl
@@ -1,4 +1,4 @@
-using CABLAB
+using ESDL
 using Base.Test
 
 include("access.jl")
@@ -6,3 +6,4 @@ include("axes.jl")
 include("analysis.jl")
 include("online.jl")
 include("artype.jl")
+include("transform.jl")